♻️ Refactor : Kanban,Todo 코드 컴포넌트 모듈화, 최적화
